More info on first commit
The first commit includes the entire project setup as well as importing in a few asset packs such as the LearninKit_Games, Animation Starter pack, and Military Weapons SIlver.

Since Github was down, this also includes retargeting animations and having them setup so I can get animations from Mixamo and have them already be retargeted to the character.

The multiplayer plugin was also added in and setup as well.