Combat Component and Weapon
Created a Combat Component for the character and added a Right Hand Socket onto the skeletal mesh for the character so that the character has a place to hold picked up weapons, and so they're positioned properly on them.
Also added in functionality for picking up weapons for the player with new input with the 'E' key.
Made it so that the weapon enum is changed when it is picked up and both its overlap sphere's collision is set to no collision and  its pickup widget is set to not show up anymore.