Leg Rotation, Server Tick changes, and Crouch Toggle
Added in rotations for the legs when the player is moving forward and left and backward and right so it no longer looks like the player is moonwalking in those directions.

Changed the server tick values to common fps game values

Changed the crouch from toggle to held like how aiming is currentlty.